Changed cube sample and line loop to double

+++ isis/src/base/objs/Cube/Cube.cpp
@@ -2135,9 +2135,9 @@ namespace Isis {
 
 
 /**
- * Returns the latitude and longitude range for the Cube. More accurate than the minimum and 
- * maximum latitude and longitude from the mapping group. 
- * 
+ * Returns the latitude and longitude range for the Cube. More accurate than the minimum and
+ * maximum latitude and longitude from the mapping group.
+ *
  * @param[out] minLatitude minimum latitude present in the cube
  * @param[out] maxLatitude maximum latitude present in the cube
  * @param[out] minLongitude minimum longitude present in the cube
@@ -2181,16 +2181,16 @@ namespace Isis {
     maxLatitude = -99999;
     maxLongitude = -99999;
 
-    for (int sample = 0.5; sample < sampleCount() + 0.5; sample++) {
+    for (double sample = 0.5; sample < sampleCount() + 0.5; sample++) {
     // Checks to see if the point is in outer space
-      for (int line = 0.5; line < lineCount() + 0.5; line++) {
+      for (double line = 0.5; line < lineCount() + 0.5; line++) {
         if (useProj) {
           isGood = proj->SetWorld(sample, line);
         }
         else {
           isGood = cam->SetImage(sample, line);
         }
-      
+
         double lat, lon;
         if (isGood) {
           if (useProj) {
@@ -2198,10 +2198,10 @@ namespace Isis {
             lon = proj->UniversalLongitude();
           }
           else {
-            lat = cam->UniversalLatitude(); 
-            lon = cam->UniversalLongitude(); 
+            lat = cam->UniversalLatitude();
+            lon = cam->UniversalLongitude();
           }
-        
+
           // update mix/max lat/lons
           if (lat < minLatitude) {
             minLatitude = lat;
@@ -2209,23 +2209,23 @@ namespace Isis {
           else if (lat > maxLatitude) {
             maxLatitude = lat;
           }
-          
+
           if (lon < minLongitude) {
             minLongitude = lon;
           }
           else if (lon > maxLongitude) {
-            maxLongitude = lon; 
+            maxLongitude = lon;
           }
         }
       }
     }
-    if ( (minLatitude == 99999) || (minLongitude == 99999) || (maxLatitude == -99999) || 
+    if ( (minLatitude == 99999) || (minLongitude == 99999) || (maxLatitude == -99999) ||
     (maxLongitude == -99999) ) {
       QString msg = "Unable to calculate a minimum or maximum latitutde or longitude.";
         throw IException(IException::Unknown, msg, _FILEINFO_);
     }
   }
-  
+
   /**
    * Write the Pvl labels to the cube's label file. Excess data in the attached
    *   labels is set to 0.
